+++ HelpOut.GitHubWorkflow.PSDevOps.ps1
@@ -12,9 +12,8 @@ New-GitHubWorkflow -Name "Test, Tag, Release, and Publish" -On Demand, Push -Job
     TagReleaseAndPublish, 
     BuildHelpOut -OutputPath .\.github\workflows\TestReleaseAndPublish.yml
 
-New-GitHubWorkflow -On Issue, 
-    Demand, 
-    Released -Job RunGitPub -Name OnIssueOrRelease -OutputPath .\.github\workflows\OnIssue.yml
+New-GitHubWorkflow -On Demand, 
+    Released -Job RunGitPub -Name OnIssueOrRelease -OutputPath .\.github\workflows\GitPub.yml
 
 
 Pop-Location